+++ src/app/angular-fundamentals/angular-fundamentals.component.html
@@ -1,0 +1,9 @@
+<h1>{{'angular - fundamentals' | titlecase}} </h1>
+<li> <a [routerLink]="['view-child']">@viewChild Decorator </a></li>
+<li> <a [routerLink]="['/']">NOT USED  </a></li>
+<li> <a [routerLink]="['/']">NOT USED   </a></li>
+
+
+
+
+<router-outlet></router-outlet>
