+++ src/app/angular-fundamentals/angular-fundamentals.component.ts
@@ -1,0 +1,19 @@
+import {Component, ElementRef, OnInit, ViewChild, AfterViewInit} from '@angular/core';
+
+@Component({
+  selector: 'app-angular-fundamentals',
+  templateUrl: './angular-fundamentals.component.html',
+  styleUrls: ['./angular-fundamentals.component.css']
+})
+export class AngularFundamentalsComponent implements OnInit, AfterViewInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+  ngAfterViewInit(){
+
+  }
+
+
+}
