+++ src/app/angular-fundamentals/components/view-child-decorator/color-picker-viewchild/color-picker-viewchild.component.css
@@ -1,4 +1,4 @@
-@import "~@angular/material/prebuilt-themes/indigo-pink.css";
+@import "../../../../../../node_modules/@angular/material/prebuilt-themes/indigo-pink.css";
 
 
 .container {
