+++ src/app/angular-fundamentals/components/view-child-decorator/color-picker-viewchild/color-picker-viewchild.component.html
@@ -1,0 +1,40 @@
+
+<h2>Their example </h2>
+<h5>This is example of the tutorial : https://blog.angular-university.io/angular-viewchild/ <br/>
+U can see here more advanced things. </h5>
+<!--clean-->
+<h3>Type injected and "read" property </h3>
+<p class = "tip1">By the read property U can defined what element U want to get , the DOM element or component if U want to.
+  <br/>On default if it's component the viewChild won't inject the DOM , but the component. </p>
+<p>U can refer an inner directive of component template ref, but U need as above to read the directive, not the element itself.<br/>
+  See at the openDialog() function the not-working-example. </p>
+
+
+<div class="container">
+
+  <h2>Choose Brand Colors:</h2>
+
+  <div class="form-field">
+
+    <color-sample
+      [color]="primary"
+      #primaryColorSample
+      (click)="openColorPicker()">
+
+    </color-sample>
+
+    <mat-input-container>
+
+      <mat-label>Primary Color</mat-label>
+
+      <input #primaryInput  name="primary"
+             [(colorPicker)]="primary" matInput
+             cpPosition="bottom"
+             [(ngModel)]="primary"/>
+
+    </mat-input-container>
+
+
+
+  </div>
+</div>
