+++ src/app/angular-fundamentals/components/view-child-decorator/color-picker-viewchild/color-picker-viewchild.component.spec.ts
