+++ src/app/angular-fundamentals/components/view-child-decorator/color-sample/color-sample.component.css
