+++ src/app/angular-fundamentals/components/view-child-decorator/view-child-decorator.component.css
@@ -1,0 +1,5 @@
+.summary1{
+  border: dodgerblue 2px solid;
+  width: auto;
+  border-radius: 5px;
+}
