+++ src/app/app.component.html
@@ -19,6 +19,10 @@
     <h2><a target="_blank" rel="noopener" href="https://blog.angular.io/">Angular blog</a></h2>
   </li>
   <li>
+    <h2><a [routerLink]="['/ang-fund']">Angular-Fundamentals</a></h2>
+
+  </li>
+  <li>
     <h2><a [routerLink]="['/learn-rxjs']">Practice RxJS</a></h2>
     <ul>
       <li>
