+++ src/app/app.module.ts
@@ -40,9 +40,11 @@ import { InnerTestComponent } from './exercises-results/inner-test/inner-test.co
 import { PagesNavigatorComponent } from './exercises-results/nba-feature/pages-navigator/pages-navigator.component';
 import { Comp1Component } from './draft-area/comp1/comp1.component';
 import {Pane, ViewChildComp} from "./draft-area/online-example/example1";
-import { ColorPickerViewchildComponent } from './draft-area/color-picker-viewchild/color-picker-viewchild.component';
-import {ColorSampleComponent} from "./draft-area/color-sample/color-sample.component";
+import { ColorPickerViewchildComponent } from './angular-fundamentals/components/view-child-decorator/color-picker-viewchild/color-picker-viewchild.component';
+import {ColorSampleComponent} from "./angular-fundamentals/components/view-child-decorator/color-sample/color-sample.component";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
+import { AngularFundamentalsComponent } from './angular-fundamentals/angular-fundamentals.component';
+import { ViewChildDecoratorComponent } from './angular-fundamentals/components/view-child-decorator/view-child-decorator.component';
 registerLocaleData(localeHe, 'lin-he');//Here u register the locale U've imported.
 registerLocaleData(localFr);//default is fr
 
@@ -81,7 +83,9 @@ registerLocaleData(localFr);//default is fr
     ViewChildComp,
     Pane,
     ColorPickerViewchildComponent,
-    ColorSampleComponent
+    ColorSampleComponent,
+    AngularFundamentalsComponent,
+    ViewChildDecoratorComponent
   ],
   imports: [
     BrowserModule,
