--- src/app/draft-area/color-picker-viewchild/color-picker-viewchild.component.html
@@ -1,63 +1,0 @@
-<div class="container">
-  <h5>This is example of the tutorial : https://blog.angular-university.io/angular-viewchild/</h5>
-  <h2>Choose Brand Colors:</h2>
-
-  <div class="form-field">
-
-    <color-sample
-      [color]="primary"
-      #primaryColorSample
-      (click)="openColorPicker()">
-
-    </color-sample>
-
-    <mat-input-container>
-
-      <mat-label>Primary Color</mat-label>
-
-      <input #primaryInput  name="primary"
-             [(colorPicker)]="primary" matInput
-             cpPosition="bottom"
-             [(ngModel)]="primary"/>
-
-    </mat-input-container>
-
-
-
-  </div>
-  <h5>The view child can defined by type, then all types will be refernced to that. Or by name... </h5>
-  <p>Angular will fill in this property automatically, but only later in the component lifecycle, after the view initialization is completed.<br/>
-    If we want to write component initialization code that uses the references injected by @ViewChild, we need to do it inside the AfterViewInit lifecycle hook.
-  </p>
-  <p>The @ViewChild decorator cannot see across component boundaries! - this mean inside inner components. </p>
-  <p>To summarize: the @ViewChild decorator is a template querying mechanism that is local to the component.
-
-    With this, we have covered the most common use case of @ViewChild, but there is still a lot more to it: let's see some more use cases!</p>
-  <color-sample
-    [color]="primary"
-    #primaryColorSample2
-    (click)="openColorPicker()">
-
-  </color-sample>
-  <h3>viewChild - simple example - </h3>
-  <p><i>See the row below that connected to the row below it. And the row below it change the p value by JS code, and the value
-  need to be known from the element. <br/>
-    Becasue of that, if you will remove (comment ) the viewchild - @ViewChild('regJsAccess'... ) it wont work</i> </p>
-  <p>the value of the below p = {{showMe}} </p>
-  <p id ="regJs" #regJsAccess>value =  initial value</p>
-  <p>See below not like that case that U don't need viewchild. That's all that they want to say. <br/>
-  When You use #refer1 inside tag, U can access that from the template itself but not from the class. </p>
-  <b #htmlRef >Case that don't viewChild </b>
-  <p>Can I access the above values ? What is it's tag ?? {{htmlRef.tagName}}</p>
-
-  <h2 #localHtml1 id = "localHtml1Id" class ="localClass1">Click me and shoot js function, that defined at the componet</h2>
-  <p>Explanation: At the above, instead of using "document.getE...('localHtml1Id').addEvent... , I am using
-  localHtml1.nativeElement.addEventListenr... . SO I can refer the element by the  class variable, not need to get it from document</p>
-  <p>By the read property U can defined what element U want to get , the DOM element or component if U want to.
-  <br/>On default if it's component the viewChild won't inject the DOM , but the component. </p>
-  <p>U can refer an inner directive of component template ref, but U need as above to read the directive, not the element itself.<br/>
-  See at the openDialog() function the not-working-example. </p>
-
-  <h2 #localHtml2 id = "localHtml2Id" class ="localClass2">initial value</h2>
-
-</div>
